Refactor warden initializer for zeitwerk
In CI, warden initializer failed on requiring the password strategy,
possibly because the module namespace did not match the file path as
zeitwerk would expect.
